Created the first version of the postprocessor
The purpose of the postprocessing module (hereafter called postprocessor) is to address issues that could not be easily handled by the mapper, which at this stage  supports only xpath expressions. Examples of issues are changing date formats, mapping facet values to closed vocabulary (for example, mapping 'en' and 'eng' to 'English').